All issues observed during the initial happy path smoke test have been resolved

+++ src/main/java/com/brokage/challenge/dto/AssetResponse.java
@@ -1,6 +1,4 @@
 package com.brokage.challenge.dto;
 
-import java.math.BigDecimal;
-
-public record AssetResponse (String customerId, String assetName, BigDecimal totalSize, BigDecimal usableSize) {
+public record AssetResponse (String customerId, String assetName, Long totalSize, Long usableSize) {
 }
